Debug

+++ API-Paris/src/main/java/com/paris/api/services/ParieService.java
@@ -19,31 +19,35 @@ public class ParieService implements Serializable {
         ParieService.repository = repository;
     }
 
-    public ParieModel findByID(Long id) {
-        return repository.findById(id).orElse(new ParieModel(null, null, 0,0,0));
-    }
-    @GetMapping("/all")
+
     public static List<ParieModel> all() {
         return repository.findAll();
+    }
+    public ParieModel findByID(Long id) {
+        return repository.findById(id).orElse(new ParieModel(null, null, 0,0,0));
     }
 
     public ParieModel createParie(ParieModel model) {
         return repository.save(model);
     }
 
-    public String deleteByID(Long id){
-        repository.deleteById(id);
-        return "Parieur supprimer";
+    public boolean delByID(Long id){
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+            return true;
+        }else{
+            return false;
+        }
+
     }
 
-    public Boolean addParie(Long idMatch, Float scoreEqip1, Float scoreEqip2){
+    public ParieModel addParie(Long idMatch, Float scoreEqip1, Float scoreEqip2){
         ParieModel newParie = new ParieModel(null, null, 0,0,0);
         newParie.setIdMatch(idMatch);
         newParie.setCoteA(calculChances(scoreEqip1, scoreEqip2, "A"));
         newParie.setCoteB(calculChances(scoreEqip1, scoreEqip2, "B"));
         newParie.setCoteN(calculChances(scoreEqip1, scoreEqip2, "N"));
-        repository.save(newParie);
-        return true;
+        return repository.save(newParie);
     }
 
     private static Float calculChances (Float score1, Float score2, String call) {
